% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prescriptive.R
\name{prescriptive}
\alias{prescriptive}
\title{Prescriptive Disparity Analysis}
\usage{
prescriptive(
  data_list,
  x,
  a,
  z,
  r,
  g,
  y,
  weight.var = NULL,
  estimator = "ri",
  K = 5,
  B = 250
)
}
\arguments{
\item{data_list}{A list of data frames, each representing an imputed dataset, or a single data frame for non-imputed data.}

\item{x}{A vector of characters corresponding to variable names.}

\item{a}{The treatment variable (a character string).}

\item{z}{The variable (a character string) for the conditional equalization intervention.}

\item{r}{The group indicator variable (a character string). A value of 1 should correspond to the disadvantaged group.}

\item{g}{The non-focal grouping variable for the affirmative action-type intervention (a character string).}

\item{y}{The outcome variable (a character string).}

\item{weight.var}{(Optional) The name of the survey weight variable (a character string).
If not provided, all observations receive equal weight.}

\item{estimator}{A character string specifying the estimator to use. Options are "ri" (regression imputation),
"w" (weighting), "dr" (doubly robust estimation), or "dml" (double machine learning). Default is "ri".}

\item{K}{For DML estimation, the number of cross-fitting folds (default is 5).}

\item{B}{For parametric estimation, the number of bootstrap iterations (default is 250).}
}
\value{
A list containing 3 data frames with the results: interventional_disparity (Eq. 19 in Opacic Wei and Zhou 2025), gap_reduced (estimates of gaps reduced), and counterfactual_means (estimates of counterfactual means under the intervention).
}
\description{
This function implements a presriptive analysis of group differences in an outcome, as proposed in Opacic Wei and Zhou (2025).
}
\examples{
\dontrun{
  # Prescriptive analysis using parametric (ri) estimator
  res_ri <- prescriptive(data_list = data, x = c("age", "female", "parinc"), a = "selective", z = "gpa", r = "black", g = "parinc", y = "comp",
                         estimator = "ri", B=250)

  # Prescriptive analysis using DML estimator
  res_dml <- prescriptive(data_list = data, x = c("age", "female", "parinc"), a = "selective", z = "gpa", r = "black", g = "parinc", y = "comp",
                          estimator = "dml", K = 5)
}
}
